# raffleease-openapi.yaml
openapi: 3.1.0
info:
  title: Raffleease API
  description: |
    Raffleease is a RESTful API designed to support the management of online raffles for animal rescue associations. The platform follows a **multi-tenant architecture**, where most endpoints are scoped under a specific association ID to ensure data isolation across organizations.
    
    The API implements **JWT-based authentication** with secure refresh token handling via **HTTP-only, same-site, secure cookies**. **Role-based access control** governs permissions across different user types within each association.
    
    Search endpoints support **advanced filtering, sorting, and pagination** mechanisms to allow flexible querying of resources. All API responses follow a **consistent, standardized format**, and **error handling** is uniform across the platform to provide clarity for integrators.
    
    This documentation should clearly outline available endpoints, authentication flows, security requirements, and expected request and response formats to help developers integrate with the Raffleease platform effectively.    

    ## Key Features
    
    ### Multi-Tenant Architecture
    Most endpoints are scoped under `/associations/{associationId}` to ensure data isolation and support multiple organizations.
    
    ### JWT-Based Authentication
    The API uses JWT access tokens for authentication and secure HTTP-only cookies to manage refresh tokens.
    
    ### Role-Based Access Control
...
    
...
    
...
    
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /v1/associations/{associationId}/images:
    post:
      tags:
        - Images
      summary: Upload pending images for raffle creation
      description: |
        Uploads images during the raffle creation process when the raffle doesn't exist yet.
        
        **Usage Context:**
        - Used during raffle creation workflow
        - Images are temporarily associated with the authenticated user and association
        - Images have "PENDING" status until the raffle is created and they are associated with it
        
        **Business Rules:**
        - Minimum 1 image, maximum 10 images per upload
        - Only ADMIN and MEMBER users can upload images
        - Images are stored temporarily and will be associated with raffle upon creation completion
        - Total images (pending + existing) cannot exceed 10 per raffle
        
        **File Requirements:**
        - Supported formats: JPEG, PNG, WebP
        - Maximum file size: 25MB per file
        - Maximum request size: 100MB total
      operationId: uploadPendingImages
      parameters:
        - $ref: '#/components/parameters/associationId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ImageUploadRequest'
            encoding:
              files:
                style: form
                explode: true
      responses:
        '200':
          description: Images uploaded successfully
          headers:
            X-Rate-Limit-Remaining:
              $ref: '#/components/headers/X-Rate-Limit-Remaining'
            X-Rate-Limit-Reset:
              $ref: '#/components/headers/X-Rate-Limit-Reset'
            X-Rate-Limit-Limit:
              $ref: '#/components/headers/X-Rate-Limit-Limit'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ImageResponse'
              example:
                success: true
                message: "New images created successfully"
                timestamp: "2024-01-15T10:30:00Z"
                data:
                  images:
                    - id: 123
                      fileName: "rescue-dog.jpg"
                      filePath: "/images/associations/456/temp/123.jpg"
                      contentType: "image/jpeg"
                      url: "https://api.raffleease.es/v1/public/associations/456/images/123"
                      imageOrder: 0
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          description: Payload Too Large - File size exceeds limits
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "File size exceeds maximum allowed limit"
                timestamp: "2024-01-15T10:30:00Z"
                statusCode: 413
                statusText: "Payload Too Large"
                code: "FILE_STORAGE_ERROR"
        '415':
          description: Unsupported Media Type - Invalid file format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "Unsupported file format. Only JPEG, PNG, and WebP are allowed"
                timestamp: "2024-01-15T10:30:00Z"
                statusCode: 415
                statusText: "Unsupported Media Type"
                code: "FILE_STORAGE_ERROR"
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/associations/{associationId}/images/images:
    get:
      tags:
        - Images
      summary: Get user's pending images
      description: |
        Retrieves all pending images uploaded by the authenticated user for the association.
        
        **Usage Context:**
        - Used during raffle creation to display uploaded images
        - Returns only images with "PENDING" status belonging to the authenticated user
        - Useful for showing preview of images before raffle creation completion
      operationId: getUserPendingImages
      parameters:
        - $ref: '#/components/parameters/associationId'
      responses:
        '200':
          description: Pending images retrieved successfully
          headers:
            X-Rate-Limit-Remaining:
              $ref: '#/components/headers/X-Rate-Limit-Remaining'
            X-Rate-Limit-Reset:
              $ref: '#/components/headers/X-Rate-Limit-Reset'
            X-Rate-Limit-Limit:
              $ref: '#/components/headers/X-Rate-Limit-Limit'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserImagesResponse'
              example:
                success: true
                message: "Images retrieved successfully"
                timestamp: "2024-01-15T10:30:00Z"
                data:
                  images:
                    - id: 123
                      fileName: "rescue-dog.jpg"
                      filePath: "/images/associations/456/temp/123.jpg"
                      contentType: "image/jpeg"
                      url: "https://api.raffleease.es/v1/public/associations/456/images/123"
                      imageOrder: 0
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/associations/{associationId}/images/{imageId}:
    delete:
      tags:
        - Images
      summary: Delete pending image
      description: |
        Soft deletes a pending image uploaded by the authenticated user.
        
        **Business Rules:**
        - Only the user who uploaded the image can delete it
        - Only images with "PENDING" status can be deleted via this endpoint
        - Performs soft deletion (marks for deletion rather than immediate removal)
        - Physical file cleanup happens during scheduled maintenance
      operationId: deletePendingImage
      parameters:
        - $ref: '#/components/parameters/associationId'
        - name: imageId
          in: path
          required: true
          description: The unique identifier of the image to delete
          schema:
            type: integer
            format: int64
            minimum: 1
          example: 123
      responses:
        '204':
          description: Image deleted successfully
          headers:
            X-Rate-Limit-Remaining:
              $ref: '#/components/headers/X-Rate-Limit-Remaining'
            X-Rate-Limit-Reset:
              $ref: '#/components/headers/X-Rate-Limit-Reset'
            X-Rate-Limit-Limit:
              $ref: '#/components/headers/X-Rate-Limit-Limit'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /v1/associations/{associationId}/raffles/{raffleId}/images/{imageId}:
    delete:
      tags:
        - Images
      summary: Delete raffle image
      description: |
        Soft deletes an image associated with a raffle.
        
        **Business Rules:**
        - Only the user who uploaded the image can delete it (if it's pending)
        - ADMIN and MEMBER users can delete any raffle images
        - Performs soft deletion (marks for deletion rather than immediate removal)
        - Physical file cleanup happens during scheduled maintenance
      operationId: deleteRaffleImage
      parameters:
        - $ref: '#/components/parameters/associationId'
        - name: raffleId
          in: path
          required: true
          description: The unique identifier of the raffle
          schema:
            type: integer
            format: int64
            minimum: 1
          example: 789
        - name: imageId
          in: path
          required: true
          description: The unique identifier of the image to delete
          schema:
            type: integer
            format: int64
            minimum: 1
          example: 124
      responses:
        '204':
          description: Image deleted successfully
          headers:
            X-Rate-Limit-Remaining:
              $ref: '#/components/headers/X-Rate-Limit-Remaining'
            X-Rate-Limit-Reset:
              $ref: '#/components/headers/X-Rate-Limit-Reset'
            X-Rate-Limit-Limit:
              $ref: '#/components/headers/X-Rate-Limit-Limit'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Public Image Serving Endpoints
  /v1/public/associations/{associationId}/images/{imageId}:
    get:
      tags:
        - Images
      summary: Get pending image file
      description: |
        Serves the actual image file for a pending image.
        
        **Usage Context:**
        - Public endpoint for displaying pending images in client applications
        - No authentication required
        - Returns the binary image file with appropriate content-type headers
        - Used for images that are uploaded during raffle creation but not yet associated with a raffle
      operationId: getPendingImageFile
      security: []  # Public endpoint
      parameters:
        - $ref: '#/components/parameters/associationId'
        - name: imageId
          in: path
          required: true
          description: The unique identifier of the image file to retrieve
          schema:
            type: integer
            format: int64
            minimum: 1
          example: 123
      responses:
        '200':
          description: Image file retrieved successfully
          headers:
            X-Rate-Limit-Remaining:
              $ref: '#/components/headers/X-Rate-Limit-Remaining'
            X-Rate-Limit-Reset:
              $ref: '#/components/headers/X-Rate-Limit-Reset'
            X-Rate-Limit-Limit:
              $ref: '#/components/headers/X-Rate-Limit-Limit'
            Content-Type:
              description: MIME type of the image
              schema:
                type: string
                enum:
                  - image/jpeg
                  - image/png
                  - image/webp
              example: "image/jpeg"
            Content-Length:
              description: Size of the image file in bytes
              schema:
                type: integer
              example: 1048576
            Cache-Control:
              description: Cache control directives
              schema:
                type: string
              example: "public, max-age=86400"
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        '404':
          description: Image not found or marked for deletion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "Image not found"
                timestamp: "2024-01-15T10:30:00Z"
                statusCode: 404
                statusText: "Not Found"
                code: "NOT_FOUND"
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/public/associations/{associationId}/raffles/{raffleId}/images/{imageId}:
    get:
      tags:
        - Images
      summary: Get raffle image file
      description: |
        Serves the actual image file for a raffle image.
        
        **Usage Context:**
        - Public endpoint for displaying raffle images in client applications
        - No authentication required
        - Returns the binary image file with appropriate content-type headers
        - Used for images that are associated with active raffles
      operationId: getRaffleImageFile
      security: []  # Public endpoint
      parameters:
        - $ref: '#/components/parameters/associationId'
        - name: raffleId
          in: path
          required: true
          description: The unique identifier of the raffle (for URL structure, not used in logic)
          schema:
            type: integer
            format: int64
            minimum: 1
          example: 789
        - name: imageId
          in: path
          required: true
          description: The unique identifier of the image file to retrieve
          schema:
            type: integer
            format: int64
            minimum: 1
          example: 124
      responses:
        '200':
          description: Image file retrieved successfully
          headers:
            X-Rate-Limit-Remaining:
              $ref: '#/components/headers/X-Rate-Limit-Remaining'
            X-Rate-Limit-Reset:
              $ref: '#/components/headers/X-Rate-Limit-Reset'
            X-Rate-Limit-Limit:
              $ref: '#/components/headers/X-Rate-Limit-Limit'
            Content-Type:
              description: MIME type of the image
              schema:
                type: string
                enum:
                  - image/jpeg
                  - image/png
                  - image/webp
              example: "image/jpeg"
            Content-Length:
              description: Size of the image file in bytes
              schema:
                type: integer
              example: 1048576
            Cache-Control:
              description: Cache control directives
              schema:
                type: string
              example: "public, max-age=86400"
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        '404':
          description: Image not found or marked for deletion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Raffle Management Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for API authentication
  parameters:
    associationId:
      name: associationId
      in: path
      required: true
      description: The unique identifier of the association for multi-tenant resource access
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 123
...
  headers:
    X-Rate-Limit-Remaining:
      description: Number of requests remaining in current rate limit window
      schema:
        type: integer
    X-Rate-Limit-Reset:
      description: Time when rate limit window resets (Unix timestamp)
      schema:
        type: integer
    X-Rate-Limit-Limit:
      description: Total number of requests allowed in rate limit window
      schema:
        type: integer
  schemas:
...
    ApiResponse:
      type: object
      discriminator:
        propertyName: success
        mapping:
          true: '#/components/schemas/SuccessResponse'
          false: '#/components/schemas/ErrorResponse'
      required:
        - success
        - message
        - timestamp
      properties:
        success:
          type: boolean
          description: Indicates whether the request was successful
        message:
          type: string
          description: Human-readable message describing the result
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of when the response was generated
      example:
        success: true
        message: "Operation completed successfully"
        timestamp: "2024-01-15T10:30:00Z"
    SuccessResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            success:
              type: boolean
              enum: [true]
            data:
              description: The response payload containing the requested data
              oneOf:
                - type: object
                - type: array
                - type: string
                - type: number
                - type: boolean
                - type: "null"
      example:
        success: true
        message: "Data retrieved successfully"
        timestamp: "2024-01-15T10:30:00Z"
        data:
          id: 123
          name: "Example"
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          required:
            - statusCode
            - statusText
            - code
          properties:
            success:
              type: boolean
              enum: [false]
            statusCode:
              type: integer
              description: HTTP status code
              minimum: 400
              maximum: 599
            statusText:
              type: string
              description: HTTP status text
            code:
              type: string
              description: Application-specific error code for programmatic handling
              enum:
                - NOT_FOUND
                - BAD_REQUEST
                - UNAUTHORIZED
                - ACCESS_DENIED
                - CONFLICT
                - FILE_STORAGE_ERROR
                - USER_NOT_FOUND
                - EMAIL_VERIFICATION_FAILED
                - PASSWORD_RESET_FAILED
                - MAIL_ERROR
                - INVALID_REQUEST
                - MISSING_PARAMETER
                - BUSINESS_ERROR
                - UNEXPECTED_ERROR
                - VALIDATION_ERROR
                - VALUE_ALREADY_EXISTS
                - RATE_LIMIT_EXCEEDED
                - IMAGE_LIMIT_EXCEEDED
                - EMAIL_SAME_AS_CURRENT
                - EMAIL_UPDATE_TOKEN_EXPIRED
                - EMAIL_UPDATE_TOKEN_INVALID
                - EMAIL_NO_LONGER_AVAILABLE
                - CURRENT_PASSWORD_INCORRECT
                - PASSWORD_SAME_AS_CURRENT
                - ROLE_UPDATE_SELF_DENIED
                - ROLE_UPDATE_ADMIN_DENIED
                - ADMIN_DISABLE_SELF_DENIED
                - ADMIN_CREATE_ADMIN_DENIED
                - INSUFFICIENT_TICKETS_AVAILABLE
      example:
        success: false
        message: "Resource not found"
        timestamp: "2024-01-15T10:30:00Z"
        statusCode: 404
        statusText: "Not Found"
        code: "NOT_FOUND"
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          required:
            - errors
          properties:
            code:
              type: string
              enum: [VALIDATION_ERROR]
            errors:
              type: object
              additionalProperties:
                type: string
              description: Field-specific validation errors mapped by field name
      example:
        success: false
        message: "Validation failed"
        timestamp: "2024-01-15T10:30:00Z"
        statusCode: 400
        statusText: "Bad Request"
        code: "VALIDATION_ERROR"
        errors:
          email: "INVALID_FORMAT"
          phoneNumber: "REQUIRED"

    # Pagination Schema
...
    ImageUploadRequest:
      type: object
      required:
        - files
      properties:
        files:
          type: array
          items:
            type: string
            format: binary
          minItems: 1
          maxItems: 10
          description: |
            Array of image files to upload.
            
            **File Requirements:**
            - Supported formats: JPEG, PNG, WebP
            - Maximum file size: 25MB per file
            - Maximum total request size: 100MB
            - Minimum 1 file, maximum 10 files per upload
      example:
        files: ["(binary image data)", "(binary image data)"]
    ImageResponse:
      type: object
      required:
        - images
      properties:
        images:
          type: array
          items:
            $ref: '#/components/schemas/ImageDTO'
          description: List of uploaded images with their metadata
      example:
        images:
          - id: 123
            fileName: "rescue-dog.jpg"
            filePath: "/images/associations/456/temp/123.jpg"
            contentType: "image/jpeg"
            url: "https://api.raffleease.es/v1/public/associations/456/images/123"
            imageOrder: 0
    UserImagesResponse:
      type: object
      required:
        - images
      properties:
        images:
          type: array
          items:
            $ref: '#/components/schemas/ImageDTO'
          description: List of user's images with their metadata
      example:
        images:
          - id: 123
            fileName: "rescue-dog.jpg"
            filePath: "/images/associations/456/temp/123.jpg"
            contentType: "image/jpeg"
            url: "https://api.raffleease.es/v1/public/associations/456/images/123"
            imageOrder: 0
    ImageDTO:
      type: object
      required:
        - id
        - fileName
        - filePath
        - contentType
        - url
        - imageOrder
      properties:
        id:
          type: integer
          format: int64
          description: The unique identifier of the image
          example: 123
        fileName:
          type: string
          minLength: 1
          description: The original name of the uploaded image file
          example: "rescue-dog.jpg"
        filePath:
          type: string
          minLength: 1
          description: The server path where the image file is stored
          example: "/images/associations/456/temp/123.jpg"
        contentType:
          type: string
          enum:
            - image/jpeg
            - image/png
            - image/webp
          description: The MIME type of the image file
          example: "image/jpeg"
        url:
          type: string
          format: uri
          description: The public URL to access the image file
          example: "https://api.raffleease.es/v1/public/associations/456/images/123"
        imageOrder:
          type: integer
          minimum: 0
          maximum: 10
          description: |
            The display order of the image within its collection.
            Lower values appear first. Range: 0-10
          example: 0
...
  responses:
...
    BadRequest:
      description: Bad Request - The request was invalid or malformed
      headers:
        X-Rate-Limit-Remaining:
          $ref: '#/components/headers/X-Rate-Limit-Remaining'
        X-Rate-Limit-Reset:
          $ref: '#/components/headers/X-Rate-Limit-Reset'
        X-Rate-Limit-Limit:
          $ref: '#/components/headers/X-Rate-Limit-Limit'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Invalid request parameters"
            timestamp: "2024-01-15T10:30:00Z"
            statusCode: 400
            statusText: "Bad Request"
            code: "BAD_REQUEST"
    Unauthorized:
      description: Unauthorized - Authentication required or token invalid
      headers:
        X-Rate-Limit-Remaining:
          $ref: '#/components/headers/X-Rate-Limit-Remaining'
        X-Rate-Limit-Reset:
          $ref: '#/components/headers/X-Rate-Limit-Reset'
        X-Rate-Limit-Limit:
          $ref: '#/components/headers/X-Rate-Limit-Limit'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Authentication required"
            timestamp: "2024-01-15T10:30:00Z"
            statusCode: 401
            statusText: "Unauthorized"
            code: "UNAUTHORIZED"
    Forbidden:
      description: Forbidden - Insufficient permissions to access the resource
      headers:
        X-Rate-Limit-Remaining:
          $ref: '#/components/headers/X-Rate-Limit-Remaining'
        X-Rate-Limit-Reset:
          $ref: '#/components/headers/X-Rate-Limit-Reset'
        X-Rate-Limit-Limit:
          $ref: '#/components/headers/X-Rate-Limit-Limit'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Access denied - insufficient permissions"
            timestamp: "2024-01-15T10:30:00Z"
            statusCode: 403
            statusText: "Forbidden"
            code: "ACCESS_DENIED"
    NotFound:
      description: Not Found - The requested resource was not found
      headers:
        X-Rate-Limit-Remaining:
          $ref: '#/components/headers/X-Rate-Limit-Remaining'
        X-Rate-Limit-Reset:
          $ref: '#/components/headers/X-Rate-Limit-Reset'
        X-Rate-Limit-Limit:
          $ref: '#/components/headers/X-Rate-Limit-Limit'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Resource not found"
            timestamp: "2024-01-15T10:30:00Z"
            statusCode: 404
            statusText: "Not Found"
            code: "NOT_FOUND"
...
    ValidationError:
      description: Unprocessable Entity - Validation failed
      headers:
        X-Rate-Limit-Remaining:
          $ref: '#/components/headers/X-Rate-Limit-Remaining'
        X-Rate-Limit-Reset:
          $ref: '#/components/headers/X-Rate-Limit-Reset'
        X-Rate-Limit-Limit:
          $ref: '#/components/headers/X-Rate-Limit-Limit'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
          example:
            success: false
            message: "Validation failed"
            timestamp: "2024-01-15T10:30:00Z"
            statusCode: 400
            statusText: "Bad Request"
            code: "VALIDATION_ERROR"
            errors:
              email: "INVALID_FORMAT"
              phoneNumber: "REQUIRED"
    TooManyRequests:
      description: Too Many Requests - Rate limit exceeded
      headers:
        X-Rate-Limit-Remaining:
          $ref: '#/components/headers/X-Rate-Limit-Remaining'
        X-Rate-Limit-Reset:
          $ref: '#/components/headers/X-Rate-Limit-Reset'
        X-Rate-Limit-Limit:
          $ref: '#/components/headers/X-Rate-Limit-Limit'
        Retry-After:
          description: Number of seconds to wait before retrying
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Rate limit exceeded. Please try again later."
            timestamp: "2024-01-15T10:30:00Z"
            statusCode: 429
            statusText: "Too Many Requests"
            code: "RATE_LIMIT_EXCEEDED"
    InternalServerError:
      description: Internal Server Error - An unexpected error occurred
      headers:
        X-Rate-Limit-Remaining:
          $ref: '#/components/headers/X-Rate-Limit-Remaining'
        X-Rate-Limit-Reset:
          $ref: '#/components/headers/X-Rate-Limit-Reset'
        X-Rate-Limit-Limit:
          $ref: '#/components/headers/X-Rate-Limit-Limit'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "An unexpected error occurred"
            timestamp: "2024-01-15T10:30:00Z"
            statusCode: 500
            statusText: "Internal Server Error"
            code: "UNEXPECTED_ERROR" 